# services/identity/contracts/openapi/identity.yaml
openapi: 3.0.3
info:
  title: UTS Identity API
  version: 0.1.0
  description: |
    Identity service: Users, Orgs, Memberships, RBAC (org/project scope), Invitations, Audit.
    - Gateway sẽ parse `/o/{org}/identity/**` và forward header `X-Org-ID: {org}`.
    - Mọi org-scoped endpoint yêu cầu header `X-Org-ID`.
    - Bảo mật: Bearer JWT (RS256, JWKS từ Identity).
servers:
  - url: https://api.dev.test
    description: Dev via Gateway
  - url: http://localhost:8081
    description: Identity local (bỏ Gateway khi chạy local)
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user + bootstrap org (OWNER)
      security: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created user & org
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile in org context
      parameters:
        - $ref: '#/components/parameters/OrgHeader'
        - $ref: '#/components/parameters/RequestId'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Me with membership snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /orgs:
    post:
      tags: [Orgs]
      summary: Create a new organization (caller becomes OWNER)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrgRequest' }
      responses:
        '201':
          description: Org created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '409': { $ref: '#/components/responses/Conflict' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    get:
      tags: [Orgs]
      summary: List organizations of current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Paged orgs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedOrganizations'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /orgs/{orgId}/members/{userId}:
    patch:
      tags: [Memberships]
      summary: Update roles/member_type of a user in org
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgHeader'
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMemberRequest' }
      responses:
        '200':
          description: Updated membership
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Membership' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Memberships]
      summary: Remove user from org
      description: Cannot remove the last OWNER of the org.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgHeader'
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Removed
        '422': { $ref: '#/components/responses/Unprocessable' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgHeader:
      name: X-Org-ID
      in: header
      required: true
      description: Organization slug or UUID, forwarded by Gateway.
      schema: { type: string }
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema: { type: string }
    OrgId:
      name: orgId
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
  responses:
    BadRequest:
      description: Bad request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Unauthorized
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Forbidden
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unprocessable:
      description: Unprocessable entity
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    UUID:
      type: string
      format: uuid
    Email:
      type: string
      format: email
    MemberType:
      type: string
      enum: [STAFF, PARTNER]
...
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        request_id: { type: string }
    User:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        email: { $ref: '#/components/schemas/Email' }
        disabled: { type: boolean }
    Organization:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        slug: { type: string }
        display_name: { type: string }
    Membership:
      type: object
      properties:
        user_id: { $ref: '#/components/schemas/UUID' }
        org_id: { $ref: '#/components/schemas/UUID' }
        roles:
          type: array
          items: { type: string }
        member_type: { $ref: '#/components/schemas/MemberType' }
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { $ref: '#/components/schemas/Email' }
        password: { type: string, minLength: 8 }
        org_name: { type: string }
        org_slug: { type: string }
    RegisterResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        org: { $ref: '#/components/schemas/Organization' }
    MeResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        org: { $ref: '#/components/schemas/Organization' }
        membership: { $ref: '#/components/schemas/Membership' }
    CreateOrgRequest:
      type: object
      required: [slug, display_name]
      properties:
        slug: { type: string }
        display_name: { type: string }
...
    UpdateMemberRequest:
      type: object
      properties:
        roles:
          type: array
          items: { type: string }
        member_type:
          $ref: '#/components/schemas/MemberType'
...
    PagedOrganizations:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Organization' }
        page: { type: integer }
        size: { type: integer }
        total: { type: integer }
...